+++ packages/terra-framework-docs/src/terra-dev-site/test/worklist-data-grid/WorklistDataGridWithRowSelection.test.jsx
@@ -44,9 +44,6 @@ const WorklistDataGridWithRowSelection = () => {
       columns={cols}
       rows={[...rows]}
       rowHeaderIndex={rowHeaderIndex}
-      rowHeight="50px"
-      columnWidth="50px"
-      columnHeaderHeight="50px"
       ariaLabel="Worklist Data Grid"
       hasSelectableRows={hasSelectableRows}
       onRowSelect={(rowId) => {