+++ packages/terra-framework-docs/src/terra-dev-site/test/worklist-data-grid/WorklistDataGridWithSelections.test.jsx
@@ -14,7 +14,7 @@ const WorklistDataGridWithSelections = () => {
         rows={rows}
         rowHeaderIndex={rowHeaderIndex}
         rowHeight="50px"
-        columnWidth="50px"
+        columnWidth="100px"
         columnHeaderHeight="50px"
         ariaLabel="Worklist Data Grid"
       />