+++ packages/terra-worklist-data-grid/tests/wdio/worklist-data-grid-spec.js
@@ -49,70 +49,69 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
     it('uses arrow keys to navigate', () => {
       browser.keys(['Tab']
-        .concat(new Array(5).fill('ArrowDown'))
-        .concat(new Array(6).fill('ArrowRight'))
-        .concat(new Array(3).fill('ArrowUp'))
-        .concat(new Array(2).fill('ArrowLeft')));
+        .concat(new Array(2).fill('ArrowDown'))
+        .concat(new Array(2).fill('ArrowRight'))
+        .concat(new Array(1).fill('ArrowUp'))
+        .concat(new Array(1).fill('ArrowLeft')));
 
+      browser.pause(150);
       Terra.validates.element('arrow-key-navigation', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the up arrow key cannot navigate outside the grid', () => {
       browser.keys(['Tab']
-        .concat(new Array(1).fill('ArrowDown'))
         .concat(new Array(3).fill('ArrowUp')));
 
+      browser.pause(150);
       Terra.validates.element('first-cell-in-grid-focused', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the left arrow key cannot navigate outside the grid', () => {
       browser.keys(['Tab']
-        .concat(new Array(1).fill('ArrowDown'))
-        .concat(new Array(3).fill('ArrowUp'))
-        .concat(new Array(3).fill('ArrowLeft')));
+        .concat(new Array(2).fill('ArrowLeft')));
 
+      browser.pause(150);
       Terra.validates.element('first-cell-in-grid-focused', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that Control + Home goes to first non-header cell', () => {
       browser.keys(['Tab']
-        .concat(new Array(4).fill('ArrowDown'))
-        .concat(new Array(3).fill('ArrowRight'))
+        .concat(new Array(2).fill('ArrowDown'))
+        .concat(new Array(2).fill('ArrowRight'))
         .concat(['Control', 'Home', 'Control']));
 
+      browser.pause(150);
       Terra.validates.element('first-non-column-header-cell-focused', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the Home key goes to the first cell in current row', () => {
       browser.keys(['Tab']
-        .concat(new Array(4).fill('ArrowDown'))
-        .concat(new Array(3).fill('ArrowRight'))
+        .concat(new Array(2).fill('ArrowDown'))
+        .concat(new Array(2).fill('ArrowRight'))
         .concat(['Home']));
 
+      browser.pause(150);
       Terra.validates.element('first-cell-in-row-focused', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the End key goes to the last cell in current row', () => {
-      browser.keys(['Tab']
-        .concat(new Array(4).fill('ArrowDown'))
-        .concat(new Array(3).fill('ArrowRight'))
-        .concat(['End']));
+      browser.keys(['Tab', 'ArrowDown', 'End']);
 
+      browser.pause(150);
       Terra.validates.element('last-cell-in-row-focused', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that Control + End key goes to the last cell in the grid', () => {
       browser.keys(['Tab']
-        .concat(new Array(4).fill('ArrowDown'))
-        .concat(new Array(3).fill('ArrowRight'))
         .concat(['Control', 'End', 'Control']));
 
+      browser.pause(150);
       Terra.validates.element('last-cell-in-last-row-focused', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
@@ -120,8 +119,9 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
     it('validates that the down arrow key cannot navigate outside the grid', () => {
       browser.keys(['Tab']
         .concat(['Control', 'End', 'Control'])
-        .concat(new Array(3).fill('ArrowDown')));
+        .concat(new Array(2).fill('ArrowDown')));
 
+      browser.pause(150);
       Terra.validates.element('last-cell-in-last-row-focused', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
@@ -129,8 +129,9 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
     it('validates that the right arrow key cannot navigate outside the grid', () => {
       browser.keys(['Tab']
         .concat(['Control', 'End', 'Control'])
-        .concat(new Array(3).fill('ArrowRight')));
+        .concat(new Array(2).fill('ArrowRight')));
 
+      browser.pause(150);
       Terra.validates.element('last-cell-in-last-row-focused', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
@@ -303,16 +304,10 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
     it('unselects a selected row by clicking in that row', () => {
       clickCell(3, 2);
-      clickCell(3, 6);
+      clickCell(3, 4);
 
       Terra.validates.element('no-rows-selected-after-click-row-selection-mode', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
-    });
-
-    it('adds correct styles to masked cells', () => {
-      browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/worklist-data-grid-masked-cell');
-
-      Terra.validates.element('masked-cell', '#terra-worklist-data-grid-masked-cell');
     });
 
     it('hovers over selectable cell in worklist data grid', () => {