+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -5,10 +5,13 @@ Added
  * Added additional keyboard navigation in `terra-worklist-data-grid`.
  * Added support for cell selection.
 
+* Changed
+  * Updated data grid to honor the column and row widths specified by consumers.
+
 ## 0.2.0 - (July 4, 2023)
 
 * Changed
-  * Removed ability to set variable row heights
+  * Removed ability to set variable row heights.
   * Updated styling on cell focus.
 * Added
   * Added masked cell support.